+++ Source/Main Dialogs/AutoTester/ORAutoTester.h
@@ -69,7 +69,7 @@
 
 #pragma mark •••Data Source Methods
 - (BOOL) outlineView:(NSOutlineView*)ov isItemExpandable:(id)item;
-- (int)  outlineView:(NSOutlineView*)ov numberOfChildrenOfItem:(id)item;
+- (NSUInteger)  outlineView:(NSOutlineView*)ov numberOfChildrenOfItem:(id)item;
 - (id)   outlineView:(NSOutlineView*)ov child:(int)index ofItem:(id)item;
 - (id)   outlineView:(NSOutlineView*)ov objectValueForTableColumn:(NSTableColumn*)tableColumn byItem:(id)item;
