+++ Source/Objects/Custom Hardware/SNO+/ResistorDB/ResistorDBModel.h
@@ -38,7 +38,7 @@
 - (void) updateResistorDb:(NSMutableDictionary*)aResistorDocDic;
 - (void) addNeweResistorDoc:(NSMutableDictionary*)aResistorDocDic;
 - (void) checkIfDocumentExists:(int)aCrate withCard:(int)aCard withChannel:(int)aChannel withRunRange:(NSMutableArray*)aRunRange;
-- (unsigned int) getCurrentRunNumber;
+- (unsigned long) getCurrentRunNumber;
 
 @end
