+++ Source/Objects/Hardware/Camac/C111C/crate_lib.h
@@ -179,27 +179,27 @@ extern "C" {
 SOCKET csock_connect	(char *ipAddress, int port);
 SOCKET csock_udpconnect	(char *ipAddress, int port, struct sockaddr_in *dest);
 
-int csock_close			(SOCKET sck);
+ssize_t csock_close			(SOCKET sck);
 
-int csock_canwrite		(SOCKET sck);
-int csock_canread		(SOCKET sck);
-int csock_send			(SOCKET sck, void *buffer, int size);
-int csock_sendto		(SOCKET sck, void *buffer, int size, struct sockaddr *dest);
+ssize_t csock_canwrite		(SOCKET sck);
+ssize_t csock_canread		(SOCKET sck);
+ssize_t csock_send			(SOCKET sck, void *buffer, int size);
+ssize_t csock_sendto		(SOCKET sck, void *buffer, int size, struct sockaddr *dest);
 
-int csock_recv			(SOCKET sck, void *buffer, int size);
-int csock_recvfrom		(SOCKET sck, void *buffer, int size, struct sockaddr *dest);
+ssize_t csock_recv			(SOCKET sck, void *buffer, int size);
+ssize_t csock_recvfrom		(SOCKET sck, void *buffer, int size, struct sockaddr *dest);
 
-int csock_recv_nb       (SOCKET sck, void *buffer, int size);
-int csock_recv_t		(SOCKET sck, void *buffer, int size, int timeout);
+ssize_t csock_recv_nb       (SOCKET sck, void *buffer, int size);
+ssize_t csock_recv_t		(SOCKET sck, void *buffer, int size, int timeout);
 
-int csock_recvline		(SOCKET sck, char *buffer, int size);
-int csock_recvline_f	(SOCKET sck, char *buffer, int size);
-int csock_recvline_t	(SOCKET sck, char *buffer, int size, int timeout);
+ssize_t csock_recvline		(SOCKET sck, char *buffer, int size);
+ssize_t csock_recvline_f	(SOCKET sck, char *buffer, int size);
+ssize_t csock_recvline_t	(SOCKET sck, char *buffer, int size, int timeout);
 
-int csock_sendrecvline  (SOCKET sck, char *cmd, char* response, int size);
-int csock_sendrecvline_t(SOCKET sck, char *cmd, char* response, int size, int timeout);
+ssize_t csock_sendrecvline  (SOCKET sck, char *cmd, char* response, int size);
+ssize_t csock_sendrecvline_t(SOCKET sck, char *cmd, char* response, int size, int timeout);
 
-int csock_flush			(SOCKET sck);
+ssize_t csock_flush			(SOCKET sck);
 
 ////////////////////////////////////////////
 //	Config Functions declaration
